move AttributeButtonGroup to Advanced components, reimport
