+++ packages/client/src/components/Advanced/index.tsx
@@ -1,5 +1,6 @@
+import { AttributeButtonGroup } from "./AttributeButtonGroup/AttributeButtonGroup";
 import { EmptyTag } from "./EmptyTag/EmptyTag";
 import { EntitySuggester } from "./EntitySuggester/EntitySuggester";
 import { EntityTag } from "./EntityTag/EntityTag";
 
-export { EmptyTag, EntitySuggester, EntityTag };
+export { AttributeButtonGroup, EmptyTag, EntitySuggester, EntityTag };
