+++ packages/client/src/pages/MainPage/containers/PropGroup/PropGroupStyles.tsx
@@ -68,11 +68,6 @@ export const StyledPropButtonGroup = styled.div<StyledPropButtonGroup>`
   border-radius: ${({ round }) => (round ? "8px" : "0")};
   border: ${({ border }) => (border ? "1px" : 0)} solid
     ${({ theme }) => theme.color["gray"][600]};
-
-  button:disabled,
-  button[disabled] {
-    //background-color: ${({ theme }) => theme.color["gray"][1000]};
-  }
 `;
 export const StyledFaGripVertical = styled(FaGripVertical)`
   margin-right: ${({ theme }) => theme.space[2]};
