+++ packages/client/src/pages/MainPage/containers/UserCustomizationModal/UserCustomizationModal.tsx
@@ -16,13 +16,12 @@ import {
   ModalInputWrap,
   Tag,
 } from "components";
-import { EntitySuggester } from "components/Advanced";
+import { AttributeButtonGroup, EntitySuggester } from "components/Advanced";
 import React, { useMemo, useState } from "react";
 import { FaUnlink } from "react-icons/fa";
 import { useMutation, useQuery, useQueryClient } from "react-query";
 import { OptionTypeBase, ValueType } from "react-select";
 import { toast } from "react-toastify";
-import { AttributeButtonGroup } from "../AttributeButtonGroup/AttributeButtonGroup";
 import {
   StyledRightsHeading,
   StyledRightsWrap,
